fix(configuration): Update examples with new config module syntax.
This fixes the imports of the read_configuration_file, change_json_key_value, and other configuration functions used to set json file parameters in the demos with the new module syntax. This also fixes various smaller bugs in the mpas example so it runs without error. The overall layout of the mpas example was also updated so all the parameters are defined up top so its clearer to users what values need to be modified. A subsequent commit will programmatically update the paths to make it easier, but this commit serves as a marker of a working set of demos using the new configuration module syntax.